requirements.txt
pyfrc ~=  2020.0.2
robotpy-rev ~= 2020.1.0
robotpy-ctre ~= 2020.2.0
wpilib ~= 2020.2.2.8

numpy
black

